feat(rocm): fp16 sparse MLA Triton kernels + indexer int64 fix for gfx1030
DeepSeek V4 Flash 0731 K128 now runs end-to-end on gfx1030 (4x V620).

- vllm/v1/attention/ops/rocm_rdna2_mla_sparse.py: fp16 Triton sparse MLA
  decode/prefill kernels. RDNA2 has no native bf16 v_dot2
  (v_dot2_f32_bf16 is RDNA3+), so the dot operands run in fp16
  (v_dot2_f32_f16). The CDNA bf16 kernels stay untouched in
  rocm_aiter_mla_sparse.py.
- csrc/rocm/indexer_paged_mqa_rdna2.cu: fix int32 overflow in slot_base
  (page_id * page_stride exceeds 2^31 on the strided indexer layout,
  producing a wild host-VA read and GPU memory fault once page_id > 2065).
  Added page_id bounds guard, is_cuda/shape TORCH_CHECKs, and an
  env-gated OOB debug print.
- vllm/models/deepseek_v4/amd/rocm.py: route gfx10x decode/prefill/inv-rope
  through the rdna2 module; fp16 gather workspace on gfx10x.
- vllm/models/deepseek_v4/common/ops/cache_utils.py: dequantize kernel
  stores in the output tensor's own dtype (bf16 on CDNA, fp16 on gfx10x).
- vllm/config/speculative.py: VLLM_DISABLE_DSPARK_MTP=1 skips auto-MTP.
- vllm/model_executor/layers/sparse_attn_indexer.py: env-gated MQA debug
  logging (VLLM_PAGED_MQA_DEBUG_OOB=1).

Bench (eager, TP=4, 32 prompts, 256+64): 61.87 total tok/s, 6.87 output
tok/s, full run completes without memory faults.

- What: fp16 sparse MLA Triton path + indexer overflow fix + routing
- Why: bf16 fdot2 unselectable on RDNA2; int32 slot_base overflow caused
  GPU memory faults mid-bench
- Testing: vllm bench throughput on 4x Radeon PRO V620, full 32-prompt
  run completes

+++ tests/kernels/quantization/test_mxfp4_rdna2.py
@@ -98,7 +98,8 @@ def _ue8m0_to_fp16_scale(scale_byte: torch.Tensor) -> torch.Tensor:
     Returns a float16 tensor with the same shape as ``scale_byte``.
     """
     bits = (scale_byte.to(torch.int32) - 112) << 10
-    return bits.to(torch.int32).view(torch.float16)
+    # via int16 so view() keeps element count (int32 view would split each in two)
+    return bits.to(torch.int16).view(torch.float16)
 
 
 def _make_e2m1_weights(E: int, K: int, N: int) -> torch.Tensor:
